# api/openapi.yaml
openapi: 3.0.1
info:
  title: E-commerce
  description: A basic API for working with an E-commerce application.
  version: 1.0.0
servers:
- url: /
paths:
...
  /signup:
    summary: Add a new user
    description: This path is used to add new users to the users.json file.
    post:
      tags:
      - Users
      summary: Add a new user
      description: This operation adds a new user to the list of orders found in the
        orders.json file.
      operationId: new_user
      requestBody:
        description: A new user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: Success
          content:
            text/plain; charset=utf-8:
              examples:
                Message:
                  value: Success
        "400":
          description: Invalid Argument Provided
          content:
            text/plain; charset=utf-8:
              examples:
                Error:
                  value: Invalid Argument
      x-swagger-router-controller: Users
  /updateusers/{id}:
    summary: Update the state of an order.
    description: This path is used to update the state of an order with a matching
      id in the orders.json file.
    put:
      tags:
      - Users
      summary: Update the state of an user
      description: This operation updates the `state` of an user with a matching id
        from the orders.json file.
      operationId: update_user
      parameters:
      - name: id
        in: path
        description: The id of the user.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: A new user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          description: Success
          content:
            text/plain; charset=utf-8:
              examples:
                Message:
                  value: Success
        "400":
          description: Invalid Argument Provided
          content:
            text/plain; charset=utf-8:
              examples:
                Error:
                  value: Invalid Argument
      x-swagger-router-controller: Users
  /deleteusers/{id}:
    summary: Delete an user
    description: This path is used to delete an user with a matching id from the users.json
      file.
    delete:
      tags:
      - Users
      summary: Deletes an user
      description: This operation deletes an user with a matching id from the users.json
        file
      operationId: delete_user
      parameters:
      - name: id
        in: path
        description: The id of the order.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success
          content:
            text/plain; charset=utf-8:
              examples:
                Message:
                  value: Success
        "400":
          description: Invalid Argument Provided
          content:
            text/plain; charset=utf-8:
              examples:
                Error:
                  value: Invalid Argument
      x-swagger-router-controller: Users
...
components:
  schemas:
...
    User:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
        nameUser:
          type: string
        passwordUser:
          type: string
      xml:
        name: User
...